+++ src/platform/unix.rs
@@ -94,7 +94,7 @@ impl Theme for WaylandThemeObject {
 }
 
 /// Additional methods on `EventLoop` that are specific to Unix.
-pub trait EventLoopExtUnix {
+pub trait EventLoopExt {
     /// Builds a new `EventLoops` that is forced to use X11.
     fn new_x11() -> Result<Self, XNotSupported>
         where Self: Sized;
@@ -120,7 +120,7 @@ pub trait EventLoopExtUnix {
     fn wayland_display(&self) -> Option<*mut raw::c_void>;
 }
 
-impl<T> EventLoopExtUnix for EventLoop<T> {
+impl<T> EventLoopExt for EventLoop<T> {
     #[inline]
     fn new_x11() -> Result<Self, XNotSupported> {
         LinuxEventLoop::new_x11().map(|ev|
@@ -171,7 +171,7 @@ impl<T> EventLoopExtUnix for EventLoop<T> {
 }
 
 /// Additional methods on `Window` that are specific to Unix.
-pub trait WindowExtUnix {
+pub trait WindowExt {
     /// Returns the ID of the `Window` xlib object that is used by this window.
     ///
     /// Returns `None` if the window doesn't use xlib (if it uses wayland for example).
@@ -226,7 +226,7 @@ pub trait WindowExtUnix {
     fn is_ready(&self) -> bool;
 }
 
-impl WindowExtUnix for Window {
+impl WindowExt for Window {
     #[inline]
     fn xlib_window(&self) -> Option<raw::c_ulong> {
         match self.window {
@@ -306,7 +306,7 @@ impl WindowExtUnix for Window {
 }
 
 /// Additional methods on `WindowBuilder` that are specific to Unix.
-pub trait WindowBuilderExtUnix {
+pub trait WindowBuilderExt {
     fn with_x11_visual<T>(self, visual_infos: *const T) -> WindowBuilder;
     fn with_x11_screen(self, screen_id: i32) -> WindowBuilder;
 
@@ -331,7 +331,7 @@ pub trait WindowBuilderExtUnix {
     fn with_app_id(self, app_id: String) -> WindowBuilder;
 }
 
-impl WindowBuilderExtUnix for WindowBuilder {
+impl WindowBuilderExt for WindowBuilder {
     #[inline]
     fn with_x11_visual<T>(mut self, visual_infos: *const T) -> WindowBuilder {
         self.platform_specific.visual_infos = Some(
@@ -390,12 +390,12 @@ impl WindowBuilderExtUnix for WindowBuilder {
 }
 
 /// Additional methods on `MonitorHandle` that are specific to Linux.
-pub trait MonitorHandleExtUnix {
+pub trait MonitorHandleExt {
     /// Returns the inner identifier of the monitor.
     fn native_id(&self) -> u32;
 }
 
-impl MonitorHandleExtUnix for MonitorHandle {
+impl MonitorHandleExt for MonitorHandle {
     #[inline]
     fn native_id(&self) -> u32 {
         self.inner.native_identifier()